Fixed read file bug, wrong memory allocation.
Some of the new memory resulting from reading the file was allocated under
the temporary file reading memory context even though it was being
returned to the caller.  Now all memory that is returned to the caller is
allocated under the top list context, which is provided by the caller.